fix<opions>: change BufferSize type from uint64 to int

+++ options.go
@@ -8,7 +8,7 @@ type TopicOption struct {
 	MaxBytesPerFile string
 	MaxDataFiles    uint64
 
-	BufferSize          uint64
+	BufferSize          int
 	BufferFlushInterval time.Duration
 }
